Clean up thread folders and protocols
