+++ Sources/App/BarcodeScanner/Camera/BarcodeScannerCamera.swift
