+++ Sources/App/BarcodeScanner/Camera/BarcodeScannerCameraView.swift
