+++ Sources/App/BarcodeScanner/Camera/BarcodeScannerDataModel.swift
