+++ Sources/App/BarcodeScanner/Camera/BarcodeScannerView.swift
