+++ Sources/App/BarcodeScanner/Camera/BarcodeScannerViewModel.swift
